Backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.3.2
pandas==2.1.3
python-dotenv==1.0.0